[Style] 파일 이름 변경
파일 이름을 바꾸기 위해 script들을 옮김
