--- Battle_DaHye/Battle_DaHye/ItemController.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace IPG
-{
-    internal class ItemController
-    {
-        public int number = 0;
-        public string Name;
-        public int Effect;
-        public string Desc;
-        public int Price;
-
-        public bool isWeapons = false;
-        public bool isSold = false;
-        public bool isUse = false;
-    }
-}
