+++ IPG/IPG/Script/ItemController.cs
@@ -13,7 +13,6 @@ namespace IPG
         public int Effect;
         public string Desc;
         public int Price;
-        public string ItemType;
 
         public bool isWeapons = false;
         public bool isSold = false;
