+++ IPG/IPG/Script/Start.cs
@@ -10,9 +10,9 @@ namespace IPG
             PlayerController player = new PlayerController();
             StoreController store = new StoreController(player);
             InventoryController inventory = new InventoryController(store, player);
+            Battlecontroller battleController = new Battlecontroller();
 
-
-            VillageController village = new VillageController(store, inventory, player);
+            VillageController village = new VillageController(store, inventory, player, battleController);
 
             store.SaveItem();
             village.Enter();
